Revert "Package all site under `front/pages/site/` (#7368)"
This reverts commit a8be7b40847a125e1cf3efbd776f2e5308e6b9b2.
