+++ front/components/home/ContentComponents.tsx
