+++ front/components/home/content/Product/BlogSection.tsx
@@ -1,15 +1,15 @@
 import React from "react";
 
-import { classNames } from "@app/lib/utils";
 import {
   Carousel,
   CarouselContent,
   CarouselItem,
   CarouselNext,
   CarouselPrevious,
-} from "@app/pages/site/components/Carousel";
-import { BlogBlock } from "@app/pages/site/components/ContentBlocks";
-import { Grid, H2, P } from "@app/pages/site/components/ContentComponents";
+} from "@app/components/home/Carousel";
+import { BlogBlock } from "@app/components/home/ContentBlocks";
+import { Grid, H2, P } from "@app/components/home/ContentComponents";
+import { classNames } from "@app/lib/utils";
 
 export function BlogSection() {
   return (
