+++ front/components/home/content/Product/FutureSection.tsx
@@ -2,15 +2,9 @@ import { Div3D, Hover3D } from "@dust-tt/sparkle";
 import Link from "next/link";
 import React from "react";
 
+import { ImgBlock } from "@app/components/home/ContentBlocks";
+import { A, Grid, H2, P, Strong } from "@app/components/home/ContentComponents";
 import { classNames } from "@app/lib/utils";
-import { ImgBlock } from "@app/pages/site/components/ContentBlocks";
-import {
-  A,
-  Grid,
-  H2,
-  P,
-  Strong,
-} from "@app/pages/site/components/ContentComponents";
 
 export function FutureSection() {
   return (
