+++ front/components/home/content/Product/TeamSection.tsx
@@ -1,12 +1,12 @@
 import { Div3D, Hover3D } from "@dust-tt/sparkle";
 import React from "react";
 
-import { Grid, H2, P } from "@app/pages/site/components/ContentComponents";
+import { Grid, H2, P } from "@app/components/home/ContentComponents";
 
 const defaultFlexClasses = "flex flex-col gap-8";
 
+import { ImgBlock } from "@app/components/home/ContentBlocks";
 import { classNames } from "@app/lib/utils";
-import { ImgBlock } from "@app/pages/site/components/ContentBlocks";
 
 export function TeamSection() {
   return (
