+++ front/components/home/menu/MobileNavigation.tsx
@@ -8,14 +8,10 @@ import type { LinkProps } from "next/link";
 import Link from "next/link";
 import * as React from "react";
 
+import { menuConfig } from "@app/components/home/menu/config";
+import { ScrollArea } from "@app/components/home/ScrollArea";
+import { Sheet, SheetContent, SheetTrigger } from "@app/components/home/Sheet";
 import { classNames } from "@app/lib/utils";
-import { menuConfig } from "@app/pages/site/components/menu/config";
-import { ScrollArea } from "@app/pages/site/components/ScrollArea";
-import {
-  Sheet,
-  SheetContent,
-  SheetTrigger,
-} from "@app/pages/site/components/Sheet";
 
 export function MobileNavigation() {
   const [open, setOpen] = React.useState(false);
