+++ front/pages/solutions/customer-support.tsx
@@ -3,16 +3,16 @@ import type { ReactElement } from "react";
 import {
   CarousselContentBlock,
   HeaderContentBlock,
-} from "@app/pages/site/components/ContentBlocks";
-import { Grid } from "@app/pages/site/components/ContentComponents";
-import type { LandingLayoutProps } from "@app/pages/site/components/LandingLayout";
-import LandingLayout from "@app/pages/site/components/LandingLayout";
+} from "@app/components/home/ContentBlocks";
+import { Grid } from "@app/components/home/ContentComponents";
+import type { LandingLayoutProps } from "@app/components/home/LandingLayout";
+import LandingLayout from "@app/components/home/LandingLayout";
 import {
   getParticleShapeIndexByName,
   shapeNames,
-} from "@app/pages/site/components/Particles";
-import type { SolutionSectionAssistantBlockProps } from "@app/pages/site/components/SolutionSection";
-import { SolutionSection } from "@app/pages/site/components/SolutionSection";
+} from "@app/components/home/Particles";
+import type { SolutionSectionAssistantBlockProps } from "@app/components/home/SolutionSection";
+import { SolutionSection } from "@app/components/home/SolutionSection";
 
 export async function getServerSideProps() {
   return {
@@ -248,7 +248,7 @@ export function CustomerCaroussel() {
       from={pageSettings.from}
       to={pageSettings.to}
       border="border-pink-100/60"
-      href="/site/solutions/customer-support"
+      href="/solutions/customer-support"
     />
   );
 }
