+++ front/pages/solutions/dust-platform.tsx
@@ -1,18 +1,18 @@
 import { Button, Div3D, Hover3D } from "@dust-tt/sparkle";
 import type { ReactElement } from "react";
 
-import { classNames } from "@app/lib/utils";
 import {
   HeaderContentBlock,
   ImgBlock,
-} from "@app/pages/site/components/ContentBlocks";
-import { Grid } from "@app/pages/site/components/ContentComponents";
-import type { LandingLayoutProps } from "@app/pages/site/components/LandingLayout";
-import LandingLayout from "@app/pages/site/components/LandingLayout";
+} from "@app/components/home/ContentBlocks";
+import { Grid } from "@app/components/home/ContentComponents";
+import type { LandingLayoutProps } from "@app/components/home/LandingLayout";
+import LandingLayout from "@app/components/home/LandingLayout";
 import {
   getParticleShapeIndexByName,
   shapeNames,
-} from "@app/pages/site/components/Particles";
+} from "@app/components/home/Particles";
+import { classNames } from "@app/lib/utils";
 
 export async function getServerSideProps() {
   return {
